+++ netpyne/network/pop.py
@@ -311,17 +311,23 @@ class Pop (object):
         for icoord, coord in enumerate(['x', 'y', 'z']):
             # constrain to range set by user
             if coord+'normRange' in self.tags:  # if normalized range, convert to normalized
-                self.tags[coord+'Range'] = [float(point) * getattr(sim.net.params, 'size'+coord.upper()) for point in self.tags[coord+'Range']]                
+                self.tags[coord+'Range'] = [float(point) * getattr(sim.net.params, 'size'+coord.upper()) for point in self.tags[coord+'normRange']]                
             if coord+'Range' in self.tags:  # if user provided absolute range, calculate range
                 self.tags[coord+'normRange'] = [float(point) / getattr(sim.net.params, 'size'+coord.upper()) for point in self.tags[coord+'Range']]
                 rangeLocs[icoord] = [self.tags[coord+'Range'][0], self.tags[coord+'Range'][1]] 
               
         gridSpacing = self.tags['gridSpacing']
         gridLocs = []
-        for x in np.arange(rangeLocs[0][0], rangeLocs[0][1]+1, gridSpacing):
-            for y in np.arange(rangeLocs[1][0], rangeLocs[1][1]+1, gridSpacing):
-                for z in np.arange(rangeLocs[2][0], rangeLocs[2][1]+1, gridSpacing):
-                    gridLocs.append((x, y, z))
+        if isinstance(gridSpacing, list):
+            for x in np.arange(rangeLocs[0][0], rangeLocs[0][1]+1, gridSpacing[0]):
+                for y in np.arange(rangeLocs[1][0], rangeLocs[1][1]+1, gridSpacing[1]):
+                    for z in np.arange(rangeLocs[2][0], rangeLocs[2][1]+1, gridSpacing[2]):
+                        gridLocs.append((x, y, z))
+        else: 
+            for x in np.arange(rangeLocs[0][0], rangeLocs[0][1]+1, gridSpacing):
+                for y in np.arange(rangeLocs[1][0], rangeLocs[1][1]+1, gridSpacing):
+                    for z in np.arange(rangeLocs[2][0], rangeLocs[2][1]+1, gridSpacing):
+                        gridLocs.append((x, y, z))
 
         numCells = len(gridLocs)